+++ kafka/java/order-camunda/src/main/java/io/flowing/retail/order/flow/ReserveStockItemsCommandPayload.java
@@ -3,13 +3,11 @@ package io.flowing.retail.order.flow;
 import io.flowing.retail.order.domain.OrderItem;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-import io.flowing.retail.order.domain.Workpiece;
 import lombok.Data;
 
 @Data
-public class CheckAvailableStockEventPayload {
+public class ReserveStockItemsCommandPayload {
     private String refId;
     private List<OrderItem> items;