+++ kafka/java/order-camunda/src/main/java/io/flowing/retail/order/messages/AllGoodsAvailableEventPayload.java
@@ -1,0 +1,18 @@
+package io.flowing.retail.order.messages;
+
+import io.flowing.retail.order.domain.OrderItem;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class AllGoodsAvailableEventPayload {
+
+    private String refId;
+    private List<OrderItem> items;
+
+}